+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ## [TODO]
 
+## [0.2.0] - 2025-08-24
+
+### Added
+- Generator to copy locales to `config/locales` in a Rails app
+
+### Changes
+- Fix railtie requirement issue
+
 ## [0.1.1] - 2025-08-23
 
 ### Changes